Allow setting a white list topic filter for kafka receivers
